requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
pyserial>=3.5
matplotlib>=3.5.0
scipy>=1.7.0
pyyaml>=6.0
pytest>=6.2.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=4.0.0